+++ .gitignore
@@ -1123,3 +1123,4 @@ v2out
 /mantis/0000854/c_out
 /mantis/0000854/a_out
 /mantis/0000864/a_out
+/Asn1f4/delme
