+++ Asn1f4/Resource1.resx
@@ -130,27 +130,6 @@
   <data name="run" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>Resources\run.sh;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
   </data>
-  <data name="asn1crt" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Resources\asn1crt.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
-  </data>
-  <data name="Acn" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Resources\Acn.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
-  </data>
-  <data name="asn1crt1" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Resources\asn1crt.h;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
-  </data>
-  <data name="ber" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Resources\ber.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
-  </data>
-  <data name="real" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Resources\real.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
-  </data>
-  <data name="real1" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Resources\real.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
-  </data>
-  <data name="xer" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Resources\xer.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
-  </data>
   <data name="adaasn1rtl_ads" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>Resources\adaasn1rtl.ads;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
   </data>
@@ -165,8 +144,7 @@
     <value>Resources\xer_rtl.ads;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
   </data>
 
-
-  <data name="adaasn1rtl_encoding_adb" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <data name="adaasn1rtl_encoding_adb" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>Resources\adaasn1rtl-encoding.adb;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
   </data>
   <data name="adaasn1rtl_encoding_ads" type="System.Resources.ResXFileRef, System.Windows.Forms">
@@ -191,6 +169,55 @@
     <value>Resources\adaasn1rtl-encoding-uper.ads;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
   </data>
 
+  <data name="asn1crt_h" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt.h;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_c" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_h" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding.h;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_c" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_uper_h" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding_uper.h;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_uper_c" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding_uper.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_acn_h" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding_acn.h;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_acn_c" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding_acn.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+  
+  <data name="asn1crt_encoding_ber_h" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding_ber.h;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_ber_c" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding_ber.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_xer_h" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding_xer.h;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  <data name="asn1crt_encoding_xer_c" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\asn1crt_encoding_xer.c;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;windows-1253</value>
+  </data>
+
+  
 
   
 </root>
