+++ StgC/body_c.stg
@@ -2,13 +2,16 @@
 
 
 
-printSourceFile(sFileNameWithoutExtension, arrsUnnamedVariables, arrsValueAssignments, arrsTypeAssignments, soMappingFunctionsModule) ::= <<
+printSourceFile(sFileNameWithoutExtension, arrsImportedRtlFiles, arrsUnnamedVariables, arrsValueAssignments, arrsTypeAssignments, soMappingFunctionsModule) ::= <<
 /*
 Code automatically generated by asn1scc tool
 */
 #include \<limits.h>
 #include \<string.h>
 #include \<math.h>
+
+<arrsImportedRtlFiles: {hdr|#include "<hdr>.h"}; separator="\n">
+
 #include "<sFileNameWithoutExtension>.h"
 <if(soMappingFunctionsModule)>
 #include "<soMappingFunctionsModule>.h"
