+++ StgC/header_c.stg
@@ -4,14 +4,19 @@ indentation(sStatement) ::=<<
     <sStatement>
 >>
 
-PrintHeaderFile(sFileNameWithNoExtUpperCase, arrsIncludedModules, arrsTypeAssignments, arrsValueAssignments, arrsPrototypes, arrsUtilityDefines) ::= <<
+PrintHeaderFile(sFileNameWithNoExtUpperCase, arrsIncludedModules, arrsTypeAssignments, arrsValueAssignments, arrsPrototypes, arrsUtilityDefines, bHasEncodings) ::= <<
 #ifndef GENERATED_ASN1SCC_<sFileNameWithNoExtUpperCase>_H
 #define GENERATED_ASN1SCC_<sFileNameWithNoExtUpperCase>_H
 /*
 Code automatically generated by asn1scc tool
 */
-<arrsIncludedModules:{inc| #include "<inc>.h"}; separator="\n">
 #include "asn1crt.h"
+<if(bHasEncodings)>
+#include "asn1crt_encoding.h"
+
+<endif>
+
+<arrsIncludedModules:{inc| #include "<inc>.h"}; separator="\n">
 
 #ifdef  __cplusplus
 extern "C" {
